pyproject.toml
[tool.poetry]
name = "ohte"
version = "0.1.0"
description = ""
authors = ["Joosef Abokther <49030615+Jouchef@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
invoke = "^2.2.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.1"
pytest = "^8.3.3"
coverage = "^7.6.7"
autopep8 = "^2.3.1"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
